\name{OutrankingTools-package}
\alias{OutrankingTools-package}
\alias{OutrankingTools}
\docType{package}
\title{Functions for Solving Multiple-criteria Decision-making Problems
}
\description{
The outranking methods constitute one of the most fruitful approach in the field of Multiple Criteria Decision Making (MCDM).
They main feature is to compare all feasible alternatives or actions by pair building up some binary relations, crisp or fuzzy, and then exploit in appropriate way these relations in order to obtain final recommendations.
This package contains functions to process ELECTRE methods existing in the literature. See, e.g., <http://en.wikipedia.org/wiki/ELECTRE> about the outranking approach and the foundations of ELECTRE methods.
}
\details{
\tabular{ll}{
Package: \tab OutrankingTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-22\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Michel Prombo <michel.prombo@statec.etat.lu> 

Maintainer: <michel.prombo@graduates.centraliens.net>
}
\references{
	Roy, B. (1996) Multiple Criteria Methodology for Decision Aiding, Dordrecht: Kluwer Academic.
	Roy, B. and Bouyssou, D. (1985). An example of comparison of two decision-aid models,in G. Fandel and J. Spronk (eds)
	Ballestero, E. and Romero, C. (1998) Multiple Criteria Decision Making and its Applications to Economic Problems, Boston-Dordrecht-London: Kluwer Academic.
	Vincke, P. (1992) Multi-criteria Decision-Aid, John Wiley, Chichester.
	Roy B. (1968) Classement et choix en presence de points de vue multiples (la methode Electre), Revue Francaise  d Informatique et de Recherche Operationnelle.
}
\keyword{ package }
\keyword{ELECTRE methods}
\keyword{Outranking approaches}
\keyword{preference modelling, multicriteria analysis}
\keyword{Sorting problem}



